added function jsl.randomizeArray() + put every function, object and class in their own file
