+++ src/deprecated.js
@@ -1,0 +1,50 @@
+/**
+ * 🔹 Returns the current version of JSLib 🔹 
+ * ⚠️ This function will be deprecated - use `jsl.info.version` or `jsl.info.intVersion` instead ⚠️
+ * @returns {String} version
+ * @since 1.5.0
+ * @version 1.8.0 This function will be deprecated - use `jsl.info.version` or `jsl.info.intVersion` instead
+ * @deprecated
+ */
+module.exports.version = () => require("./objects/info").version;
+
+/**
+ * 🔹 Adds color(s) to the input text and sends that colored text to the console 🔹
+ * ⚠️ This function will be deprecated - use `jsl.colors` instead ⚠️
+ * @param {String} text the text that should be colored and sent as a console message
+ * @param {String} colors space separated list of color(s). (Available colors are: "rst/reset, bright, dim, underscore/ul/underline, blink, reverse, hidden, fgblack, fgred, fggreen, fgyellow, fgblue, fgmagenta, fgcyan, fgwhite, bgblack, bgred, bggreen, bgyellow, bgblue, bgmagenta, bgcyan, bgwhite")
+ * @returns executes console.log() function
+ * @since 1.6.0
+ * @deprecated This function will soon be deprecated. Please use the object `jsl.colors` instead - This function will soon redirect to that object too and then get completely deprecated in a later version
+ */
+module.exports.consoleColor = (text, colors) => {
+    let cnbr = [];
+    if(colors.includes("rst") || colors.includes("reset")) cnbr.push(0);
+    if(colors.includes("bright")) cnbr.push(1);
+    if(colors.includes("dim")) cnbr.push(2);
+    if(colors.includes("underscore") || colors.includes("ul") || colors.includes("underline")) cnbr.push(4);
+    if(colors.includes("blink")) cnbr.push(5);
+    if(colors.includes("reverse")) cnbr.push(7);
+    if(colors.includes("hidden")) cnbr.push(8);
+
+    if(colors.includes("fgblack")) cnbr.push(30);
+    if(colors.includes("fgred")) cnbr.push(31);
+    if(colors.includes("fggreen")) cnbr.push(32);
+    if(colors.includes("fgyellow")) cnbr.push(33);
+    if(colors.includes("fgblue")) cnbr.push(34);
+    if(colors.includes("fgmagenta")) cnbr.push(35);
+    if(colors.includes("fgcyan")) cnbr.push(36);
+    if(colors.includes("fgwhite")) cnbr.push(37);
+
+    if(colors.includes("bgblack")) cnbr.push(40);
+    if(colors.includes("bgred")) cnbr.push(41);
+    if(colors.includes("bggreen")) cnbr.push(42);
+    if(colors.includes("bgyellow")) cnbr.push(43);
+    if(colors.includes("bgblue")) cnbr.push(44);
+    if(colors.includes("bgmagenta")) cnbr.push(45);
+    if(colors.includes("bgcyan")) cnbr.push(46);
+    if(colors.includes("bgwhite")) cnbr.push(47);
+
+    for(let i = 0; i < cnbr.length; i++) cnbr[i] = "\x1b[" + cnbr[i] + "m";
+    return console.log(cnbr.join("") + text + "\x1b[0m");
+}
