+++ src/functions/allEqual.js
@@ -1,0 +1,17 @@
+/**
+ * 🔹 Tests an array and returns true if all values are equal. 🔹
+ * @param {Array} array
+ * @returns {Boolean} true if all values are equal, false if not
+ * @throws Throws an error if the parameter is not an array
+ * @since 1.5.0
+ * @version 1.8.0 Throwing error now instead of returning string
+ */
+const allEqual = array => {
+    let isEmpty = require("./isEmpty");
+
+    if(isEmpty(array) || typeof array != "object") 
+        throw new Error(`Wrong arguments provided for jsl.allEqual() - (expected: "Object/Array", got: "${typeof array}")`);
+
+    return array.every(v => v === array[0]);
+}
+module.exports = allEqual;
