+++ src/functions/isEmpty.js
@@ -1,0 +1,13 @@
+/**
+ * 🔹 Returns true, if the input is undefined, null, an empty string, an empty array or an object with length = 0. 
+ * Otherwise returns false. The number 0 and NaN will return false though, so check them independently if needed! 🔹
+ * @param {*} input Variable that should be checked - this can be of any type but the basic types will work best
+ * @returns {Boolean} true or false
+ * @since 1.4.0
+ * @version 1.6.5 lowercase alias jsl.isempty was removed
+ * @version 1.8.0 Added check for objects with length = 0
+ */
+const isEmpty = input => {
+    return (input === undefined || input === null || input === "" || (typeof input == "object" && input.length < 0)) ? true : false;
+}
+module.exports = isEmpty;
