+++ src/functions/seededRNG/validateSeed.js
@@ -1,0 +1,26 @@
+/**
+ * 🔹 Validates a seed 🔹
+ * @param {(Number|String)} seed 
+ * @returns {Boolean}
+ * @since 1.8.0
+ */
+const validateSeed = (seed) => {
+    let isEmpty = require("../isEmpty");
+    let digitCount;
+    
+    if(typeof seed == "string") digitCount = parseInt(seed.length);
+    else digitCount = parseInt(seed.toString().length);
+
+    if(isEmpty(seed) || isEmpty(digitCount) || isNaN(parseInt(digitCount)))
+        throw new Error(`Invalid argument provided for validateSeed() - make sure it is not empty / null / undefined and is of the correct type.\nExpected: "number" or "string", got: "${typeof seed}"`);
+
+    seed = seed.toString();
+
+    let regex = new RegExp(`^[0-9]{${digitCount}}`, "gm");
+
+    if(!seed.match(regex) || seed.includes("\n"))
+        return false;
+
+    return true;
+}
+module.exports = validateSeed;
